function to create matrixes
the function reset draw lost his use with the last updates, so it was renamed to create matrixes in an unique place, just the calll where is needed.
Also, with generic var, in draw scramble now is able to have only one structure to pass the values matrix to color, and using for structure to interate easily.